#8517 ui part of ext.bigquery plugin moved in ext.bigquery.ui
