+++ plugins/org.jkiss.dbeaver.ext.bigquery.ui/OSGI-INF/l10n/bundle.properties
@@ -1,0 +1,4 @@
+Bundle-Vendor = DBeaver Corp
+Bundle-Name = DBeaver Google BigQuery UI
+
+bigquery.dialog.connection.header=Google BigQuery Connection Settings
