+++ plugins/org.jkiss.dbeaver.ext.bigquery.ui/OSGI-INF/l10n/bundle_de.properties
@@ -1,0 +1,1 @@
+bigquery.dialog.connection.header=Google BigQuery Verbindungseinstellungen
