+++ plugins/org.jkiss.dbeaver.ext.bigquery.ui/OSGI-INF/l10n/bundle_pt_BR.properties
@@ -1,0 +1,4 @@
+Bundle-Name = DBeaver Google BigQuery UI
+Bundle-Vendor = DBeaver Corp 
+
+bigquery.dialog.connection.header = Google BigQuery Configura\u00E7\u00F5es de conex\u00E3o
