+++ plugins/org.jkiss.dbeaver.ext.bigquery.ui/src/org/jkiss/dbeaver/ext/bigquery/ui/BigQueryActivator.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jkiss.dbeaver.ext.bigquery;
+package org.jkiss.dbeaver.ext.bigquery.ui;
 
 import org.eclipse.jface.resource.ImageDescriptor;
 import org.eclipse.ui.plugin.AbstractUIPlugin;
