+++ plugins/org.jkiss.dbeaver.ext.bigquery.ui/src/org/jkiss/dbeaver/ext/bigquery/ui/internal/messages_de.properties
@@ -1,0 +1,13 @@
+label_additional_project=Zus\u00e4tzliches Projekt(e)
+label_connection=Verbindung
+label_host=Hostrechner
+label_key_path=Schl\u00fcsselpfad
+label_oauth_type=OAuth Typ
+label_port=Port
+label_private_key_path=Privater Schl\u00fcsselpfad (p12 oder JSON)
+label_project=Projekt
+label_security=Sicherheit
+label_server_info=Serverinfo
+label_service_account=Dienstkonto
+label_service_based=Service-basiert
+label_user_based=Benutzerbasiert
