+++ plugins/org.jkiss.dbeaver.ext.bigquery/META-INF/MANIFEST.MF
@@ -4,14 +4,11 @@ Bundle-Name: DBeaver BigQuery extension
 Bundle-SymbolicName: org.jkiss.dbeaver.ext.bigquery;singleton:=true
 Bundle-Version: 1.0.37.qualifier
 Bundle-Release-Date: 20200420
-Require-Bundle: org.eclipse.ui,
- org.eclipse.core.runtime,
- org.jkiss.dbeaver.core,
- org.jkiss.dbeaver.model,
- org.jkiss.dbeaver.ext.generic
-Bundle-Activator: org.jkiss.dbeaver.ext.bigquery.BigQueryActivator
+Require-Bundle: org.jkiss.dbeaver.ext.generic
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-Vendor: DBeaver Corp
 Bundle-Localization: OSGI-INF/l10n/bundle
 Automatic-Module-Name: org.jkiss.dbeaver.ext.bigquery
+Export-Package: org.jkiss.dbeaver.ext.bigquery,
+ org.jkiss.dbeaver.ext.bigquery.model
