+++ plugins/org.jkiss.dbeaver.ext.bigquery/OSGI-INF/l10n/bundle.properties
@@ -1,5 +1,3 @@
 Bundle-Vendor = DBeaver Corp
 Bundle-Name = DBeaver Google BigQuery Extension
 datasource.bq.description = BigQuery datasource
-
-bigquery.dialog.connection.header=Google BigQuery Connection Settings
