+++ plugins/org.jkiss.dbeaver.ext.bigquery/OSGI-INF/l10n/bundle_pt_BR.properties
@@ -1,8 +1,5 @@
 
 Bundle-Name = DBeaver Google BigQuery Extens\u00E3o
-
 Bundle-Vendor = DBeaver Corp 
-
-bigquery.dialog.connection.header = Google BigQuery Configura\u00E7\u00F5es de conex\u00E3o
 
 datasource.bq.description = BigQuery Origem de dados
