+++ plugins/org.jkiss.dbeaver.ext.bigquery/OSGI-INF/l10n/bundle_ru.properties
@@ -1,1 +1,1 @@
-bigquery.dialog.connection.header=\u041D\u0430\u0441\u0442\u0440\u043E\u0439\u043A\u0438 \u043F\u043E\u0434\u043A\u043B\u044E\u0447\u0435\u043D\u0438\u044F Google BigQuery
+datasource.bq.description = BigQuery \u0438\u0441\u0442\u043e\u0447\u043d\u0438\u043a\u0020\u0434\u0430\u043d\u043d\u044b\u0445
